Adjust LineStrings in Case of V3 export correctly. Fix some formattings
