+++ Dev4Agriculture.ISO11783.ISOXML.Test/TaskDataAnalysisTests.cs
@@ -2,55 +2,55 @@
 using de.dev4Agriculture.ISOXML.DDI;
 using Dev4Agriculture.ISO11783.ISOXML.Analysis;
 using Dev4Agriculture.ISO11783.ISOXML.IdHandling;
+using Dev4Agriculture.ISO11783.ISOXML.Utils;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace Dev4Agriculture.ISO11783.ISOXML.Test
+namespace Dev4Agriculture.ISO11783.ISOXML.Test;
+
+[TestClass]
+public class TaskDataAnalysisTests
 {
-    [TestClass]
-    public class TaskDataAnalysisTests
+    [TestMethod]
+    public void CanCalculateTotals()
     {
-        [TestMethod]
-        public void CanCalculateTotals()
-        {
-            var isoxml = ISOXML.Load("./testdata/TimeLogs/TotalsTests");
-            //Testing LifeTime Totals
-            Assert.IsTrue(isoxml.Data.Task[0].TryGetTotalValue(0x114, 0, out var totalFuelLifeTime, TLGTotalAlgorithmType.LIFETIME));
-            Assert.AreEqual(totalFuelLifeTime, 1294 / 0.5);
+        var isoxml = ISOXML.Load("./testdata/TimeLogs/TotalsTests");
+        //Testing LifeTime Totals
+        Assert.IsTrue(isoxml.Data.Task[0].TryGetTotalValue(0x114, 0, out var totalFuelLifeTime, TLGTotalAlgorithmType.LIFETIME));
+        Assert.AreEqual(totalFuelLifeTime, 1294 / 0.5);
 
 
-            //Testing Task Total
-            var analysis = new ISODeviceAnalysis(isoxml);
-            //We know it's all the same DET in this case, so we only call it once for all DDIs
-            var detList = analysis.FindDeviceElementsForDDI(isoxml.Data.Task[0], 0x0078);
-            var detId = IdList.ToIntId(detList[0].DeviceElementId);
-            Assert.IsTrue(isoxml.Data.Task[0].TryGetTotalValue(0x78, detId, out var totalInEffectiveTime, TLGTotalAlgorithmType.NO_RESETS))
-            ;
-            Assert.AreEqual(totalInEffectiveTime, 4531 /*Close to 75.5 minutes*/);
+        //Testing Task Total
+        var analysis = new ISODeviceAnalysis(isoxml);
+        //We know it's all the same DET in this case, so we only call it once for all DDIs
+        var detList = analysis.FindDeviceElementsForDDI(isoxml.Data.Task[0], 0x0078);
+        var detId = IdList.ToIntId(detList[0].DeviceElementId);
+        Assert.IsTrue(isoxml.Data.Task[0].TryGetTotalValue(0x78, detId, out var totalInEffectiveTime, TLGTotalAlgorithmType.NO_RESETS))
+        ;
+        Assert.AreEqual(totalInEffectiveTime, 4531 /*Close to 75.5 minutes*/);
 
-            //Testing Task Maximum
-            Assert.IsTrue(isoxml.Data.Task[0].TryGetMaximum(0x43, detId, out var maximum));
-            Assert.AreEqual(maximum, 12000);
+        //Testing Task Maximum
+        Assert.IsTrue(isoxml.Data.Task[0].TryGetMaximum(0x43, detId, out var maximum));
+        Assert.AreEqual(maximum, 12000);
 
 
-            Assert.IsTrue(isoxml.Data.Task[0].TryGetTotalValue(0xB7, detId, out var totalDryMass,TLGTotalAlgorithmType.NO_RESETS));
-            Assert.AreEqual(totalDryMass, 2000);
+        Assert.IsTrue(isoxml.Data.Task[0].TryGetTotalValue(0xB7, detId, out var totalDryMass, TLGTotalAlgorithmType.NO_RESETS));
+        Assert.AreEqual(totalDryMass, 2000);
 
-            var timeElements = isoxml.Data.Task[0].GenerateTimeElementsFromTimeLogs(isoxml.Data.Device);
-            Assert.AreEqual(timeElements.Count, 3);
-            Assert.AreEqual(timeElements[2].DataLogValue.First(entry => Utils.ConvertDDI(entry.ProcessDataDDI) == (ushort)DDIList.IneffectiveTotalTime).ProcessDataValue, 4531);
-            Assert.AreEqual(timeElements[2].DataLogValue.First(entry => Utils.ConvertDDI(entry.ProcessDataDDI) == (ushort)DDIList.LifetimeTotalArea).ProcessDataValue, 561780);
-        }
+        var timeElements = isoxml.Data.Task[0].GenerateTimeElementsFromTimeLogs(isoxml.Data.Device);
+        Assert.AreEqual(timeElements.Count, 3);
+        Assert.AreEqual(timeElements[2].DataLogValue.First(entry => DDIUtils.ConvertDDI(entry.ProcessDataDDI) == (ushort)DDIList.IneffectiveTotalTime).ProcessDataValue, 4531);
+        Assert.AreEqual(timeElements[2].DataLogValue.First(entry => DDIUtils.ConvertDDI(entry.ProcessDataDDI) == (ushort)DDIList.LifetimeTotalArea).ProcessDataValue, 561780);
+    }
 
-        [TestMethod]
-        public void CanReadTaskExtract()
-        {
-            var isoxml = ISOXML.Load("./testdata/TimeLogs/TotalsTests");
-            //Testing LifeTime Totals
-            var entries = isoxml.Data.Task[0].GetMergedTaskExtract(0x0043, -1);
-            var filledEntries = isoxml.Data.Task[0].GetMergedTaskExtract(0x0043, -1, "", true);
-            Assert.AreEqual(entries.Data.Count, 2);
-            Assert.AreEqual(filledEntries.Data.Count, 4677);
-        }
+    [TestMethod]
+    public void CanReadTaskExtract()
+    {
+        var isoxml = ISOXML.Load("./testdata/TimeLogs/TotalsTests");
+        //Testing LifeTime Totals
+        var entries = isoxml.Data.Task[0].GetMergedTaskExtract(0x0043, -1);
+        var filledEntries = isoxml.Data.Task[0].GetMergedTaskExtract(0x0043, -1, "", true);
+        Assert.AreEqual(entries.Data.Count, 2);
+        Assert.AreEqual(filledEntries.Data.Count, 4677);
     }
 }
