+++ Dev4Agriculture.ISO11783.ISOXML/ISOExtensions/ISOAllocationStampExtension.cs
@@ -10,14 +10,14 @@ namespace Dev4Agriculture.ISO11783.ISOXML.TaskFile
 
         public ulong GetSeconds()
         {
-            if (this.Stop != null)
+            if (Stop != null)
             {
-                TimeSpan? duration = Stop - Start;
+                var duration = Stop - Start;
                 return (ulong)duration?.TotalSeconds;
             }
             else if (Duration != null)
             {
-                return Duration??0;
+                return Duration ?? 0;
             } else
             {
                 return 0;
@@ -27,9 +27,9 @@ namespace Dev4Agriculture.ISO11783.ISOXML.TaskFile
 
         public bool TryGetSeconds(out ulong seconds)
         {
-            if (this.Stop != null)
+            if (Stop != null)
             {
-                TimeSpan? duration = Stop - Start;
+                var duration = Stop - Start;
                 seconds = (ulong)duration?.TotalSeconds;
                 return true;
             }
@@ -48,7 +48,7 @@ namespace Dev4Agriculture.ISO11783.ISOXML.TaskFile
 
         internal void FixPositionDigits()
         {
-            if( Position != null )
+            if (Position != null)
             {
                 Position.ToList().ForEach( ptn => ptn.FixDigits());
             }
