+++ Dev4Agriculture.ISO11783.ISOXML/ISOExtensions/ISOTimeExtension.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using Dev4Agriculture.ISO11783.ISOXML.IdHandling;
+using Dev4Agriculture.ISO11783.ISOXML.Utils;
 
 namespace Dev4Agriculture.ISO11783.ISOXML.TaskFile
 {
@@ -11,9 +11,9 @@ namespace Dev4Agriculture.ISO11783.ISOXML.TaskFile
 
         public ulong GetSeconds()
         {
-            if (this.Stop != null)
+            if (Stop != null)
             {
-                TimeSpan? duration = Stop - Start;
+                var duration = Stop - Start;
                 return (ulong)duration?.TotalSeconds;
             }
             else if (Duration != null)
@@ -29,10 +29,10 @@ namespace Dev4Agriculture.ISO11783.ISOXML.TaskFile
 
         public bool TryGetSeconds(out ulong seconds)
         {
-            if (this.Stop != null)
+            if (Stop != null)
             {
-                TimeSpan? duration = Stop - Start;
-                seconds =  (ulong)duration?.TotalSeconds;
+                var duration = Stop - Start;
+                seconds = (ulong)duration?.TotalSeconds;
                 return true;
             }
             else if (Duration != null)
@@ -52,13 +52,13 @@ namespace Dev4Agriculture.ISO11783.ISOXML.TaskFile
             foreach (var dlv in tim.DataLogValue)
             {
                 var compare = lastTim.DataLogValue.First(entry =>
-                                                            Utils.ConvertDDI(entry.ProcessDataDDI) == Utils.ConvertDDI(dlv.ProcessDataDDI) &&
+                                                            DDIUtils.ConvertDDI(entry.ProcessDataDDI) == DDIUtils.ConvertDDI(dlv.ProcessDataDDI) &&
                                                             entry.DeviceElementIdRef == dlv.DeviceElementIdRef
                                                         );
                 if (compare != null)
                 {
                     var device = devices.First(dvc => dvc.DeviceElement.Any(det => det.DeviceElementId == dlv.DeviceElementIdRef));
-                    if (device.IsLifetimeTotal(Utils.ConvertDDI(dlv.ProcessDataDDI)))
+                    if (device.IsLifetimeTotal(DDIUtils.ConvertDDI(dlv.ProcessDataDDI)))
                     {
                         dlv.ProcessDataValue = compare.ProcessDataValue;
                     }
@@ -81,7 +81,7 @@ namespace Dev4Agriculture.ISO11783.ISOXML.TaskFile
         /// <returns>True if a value was found</returns>
         public bool TryGetDDIValue(ushort ddi, int deviceElement, out int lastValue)
         {
-            var dlv = DataLogValue.ToList().First(entry => Utils.ConvertDDI(entry.ProcessDataDDI) == ddi && IdList.ToIntId(entry.DeviceElementIdRef) == deviceElement)?.ProcessDataValue;
+            var dlv = DataLogValue.ToList().First(entry => DDIUtils.ConvertDDI(entry.ProcessDataDDI) == ddi && IdList.ToIntId(entry.DeviceElementIdRef) == deviceElement)?.ProcessDataValue;
             if (dlv != null)
             {
                 lastValue = (int)dlv;
