+++ Dev4Agriculture.ISO11783.ISOXML/ISOGridFile.cs
@@ -1,8 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.IO;
 using Dev4Agriculture.ISO11783.ISOXML.Messaging;
 using Dev4Agriculture.ISO11783.ISOXML.TaskFile;
+using Dev4Agriculture.ISO11783.ISOXML.Utils;
 
 namespace Dev4Agriculture.ISO11783.ISOXML
 {
@@ -44,7 +44,7 @@ namespace Dev4Agriculture.ISO11783.ISOXML
         {
             var result = new ResultWithMessages<ISOGridFile>();
             ISOGridFile grid = null;
-            if (Utils.AdjustFileNameToIgnoreCasing(baseFolder, name + ".bin", out var filePath))
+            if (FileUtils.AdjustFileNameToIgnoreCasing(baseFolder, name + ".bin", out var filePath))
             {
                 grid = new ISOGridFile(type);
 
@@ -65,9 +65,9 @@ namespace Dev4Agriculture.ISO11783.ISOXML
                                 {
                                     for (var x = 0; x < grid.Width; x++)
                                     {
-                                        for(var l = 0; l < grid.Layers; l++)
+                                        for (var l = 0; l < grid.Layers; l++)
                                         {
-                                            grid._datat1[y, x, l] = buffer[x*grid.Layers + l];
+                                            grid._datat1[y, x, l] = buffer[x * grid.Layers + l];
                                         }
                                     }
                                 }
@@ -159,7 +159,7 @@ namespace Dev4Agriculture.ISO11783.ISOXML
                     {
                         for (var x = 0; x < Width; x++)
                         {
-                            for(var l = 0; l < Layers; l++)
+                            for (var l = 0; l < Layers; l++)
                             {
                                 bw.Write(_datat1[y, x, l]);
                             }
@@ -192,9 +192,9 @@ namespace Dev4Agriculture.ISO11783.ISOXML
         {
             try
             {
-                for(var l=0; l<Layers; l++)
+                for (var l = 0; l < Layers; l++)
                 {
-                    var filePath = storagePath + Name + (Layers >1 ? "_" + l : "") + ".CSV";
+                    var filePath = storagePath + Name + (Layers > 1 ? "_" + l : "") + ".CSV";
                     var file = File.Create(filePath);
                     var streamWriter = new StreamWriter(file);
                     switch (Type)
@@ -271,7 +271,7 @@ namespace Dev4Agriculture.ISO11783.ISOXML
         /// <returns></returns>
         public int SetValue(uint column, uint row, int value, uint layer = 0)
         {
-            if (column < 0 || column >= Width || row < 0 || row >= Height || layer <0 || layer >= Layers)
+            if (column < 0 || column >= Width || row < 0 || row >= Height || layer < 0 || layer >= Layers)
             {
                 return 0;
             }
