+++ Dev4Agriculture.ISO11783.ISOXML/Utils/HexUtils.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Linq;
+
+namespace Dev4Agriculture.ISO11783.ISOXML.Utils
+{
+    public class HexUtils
+    {
+
+        public static byte[] HexStringToByteArray(string hex)
+        {
+            return Enumerable.Range(0, hex.Length)
+                             .Where(x => x % 2 == 0)
+                             .Select(x => Convert.ToByte(hex.Substring(x, 2), 16))
+                             .ToArray();
+        }
+
+        public static string ByteArrayToHexString(byte[] data)
+        {
+            return BitConverter.ToString(data).Replace("-", "");
+        }
+    }
+}
