+++ Dev4Agriculture.ISO11783.ISOXML/Utils/StringUtils.cs
@@ -1,0 +1,26 @@
+﻿using System.IO;
+using System.Text;
+
+namespace Dev4Agriculture.ISO11783.ISOXML.Utils
+{
+    internal class StringUtils
+    {
+    }
+
+    public sealed class StringWriterWithEncoding : StringWriter
+    {
+        private readonly Encoding encoding;
+
+        public StringWriterWithEncoding() : this(Encoding.UTF8) { }
+
+        public StringWriterWithEncoding(Encoding encoding)
+        {
+            this.encoding = encoding;
+        }
+
+        public override Encoding Encoding
+        {
+            get { return encoding; }
+        }
+    }
+}
